Use new hugo pipes to generate assets
Instead of using wellington for generating assets for the site I've
moved into using hugo pipes. This means that generating the site is even
easier than before and faster. With this change I've also introduced a
small change in the asset directory because I'm merging the site and
resume sites.